Update the comment describing the grammar.
It turns out that syntax errors produced by an odd
number of vertical bars in block comments are intended
behaviour, because block comments are parsed by the
compiler so that, for example, nested block comments
are possible. Therefore, vertical-bar quoted identifiers
must be well-formed even inside block comments.

Use line comments (possibly nested inside a block
comment as in this case) to let the compiler
completely ignore the content of the comment.