Implement some path operations using NSURL.
I have defined a `_fileURL()` handler to be used
internally to build an NSURL object from a path
specification. The following handlers have been
re-written to use `_fileURL()`:

    absolutePath()
    basename()
    directoryPath() (was dirname())
    joinPath()
    splitPath()

Note that POSIX paths are returned by these handlers
by getting the `lastPathComponent`, `path`,
`relativePath` properties of an NSURL, which strip
a trailing slash if it exists. This is something
to keep in mind for updating the tests.