Re-design the way tasks with parameters are run.
In principle, we can pass arguments to tasks via
the run handler (using `run script <task> with parameters`). Unfortunately, using `run script`
seems to inhibit logging to stdout. Try:

script Foo	on run argv		-- argv will be script Foo when no arguments are passed		if class of argv is list then			repeat with i from 1 to argv's length				log item i of argv			end repeat		end if	end runend scriptrun script Foo with parameters {"a", 1} — Ok, but no output

Besides, according to The AS Language Guide,
`run script` should not accept script objects.
Therefore, we use the following workaround:

script Foo
	property argv : {}

	on exec:(argv as list)
		set my argv to argv
		run me
	end exec:

	on run
		repeat with i from 1 to (count my argv)
			log item i of argv
		end repeat
	end run
end script

Foo's exec:{"a", 1}

The `exec` handler is defined by Task(), which now
creates a script object that is between TaskBase
and the user task in the inheritance chain.

The `arguments` property is not necessary any longer, so we remove it. The TaskArguments script
is renamed to CommandLine.