Implement handlers to build script and app bundles.
This is a major refactoring of the code with lots of
changes. The most important is that now we have
handlers to build script bundles and applications
from source, which support resources, auxiliary
scripts and script libraries in source form (which
are built recursively).

Most path and file manipulation handlers have been
split into “low-level” handlers that operate on ObjC
references (which are prefixed by an underscore and
are meant to be used only internally by ASMake) and
handlers that operate on AS entities (which may be
used by client code). The former are called by the
latter. The low-level handlers can be called directly internally, avoiding too many conversions.