Load ASMake at runtime in makefile.
Use a cleaner approach to load ASMake from source
in the makefile. Now it is loaded at runtime and
tasks are registered at runtime, too.
By doing so, I may get rid of the ugly _setpath()
handler. An another advantage is that now the
makefile can be run from a different directory
and it can be executed in compiled form.

I have also added a (pretty useless) “args” task
that simply prints out its arguments and exits.