Create a test folder for tests.
We now build the tests into a script bundle with
ASMake as an embedded script library. By doing so,
we can run the tests without installing ASMake
and we can simplify the header code in Test ASMake.applescript, getting rid of the ugly _wd()
handler.

Note that most tests are broken.