Use a “trick” in the makefile to process options.
Although it is not possible to pass the `run`
commands any arguments, the form `continue run`
allows you to specify arguments. To use it,
we use a small trick: we define a script that
inherits from ASMake and then we “continue run” in
that script, so that the parent’s run handler is
called.

Maybe we could use a similar trick in `exec:`, too.