Fix API doc generation.
HeaderDoc’s ability to process AppleScript scripts
has improved a lot with Yosemite (thanks to my bug
reports, too!), but it is not very robust yet.
HeaderDoc needs some help to generate a complete
documentation. Often, problems are resolved by
doing one or more of the following:

- property values split on several lines;
- conditionals on a single line (if… then…);
- tell… to… statements on a single line.

In some cases HeaderDoc may still skip a handler for
reasons that are still obscure to me. In such cases, exchanging the order of the handlers may solve the problem.
This is the case for `run argv`: putting before
`runTask()` makes HeaderDoc recognize it and
generate an entry in the documentation.

The changes in this commit are the minimal changes
needed to have the documentation produced correctly.
Note that applying the rules mentioned above
systematically may in fact confuse HeaderDoc even
more.