Prevent automatic execution of tests.
Most likely because of bugs in the current version of
AppleScript (in the `load script` and `run script`
commands), I cannot find a way to run the test suite
from the makefile. The method used before this
commit (load the compiled test suite with `load script`
then run it) causes a “Segmentation fault: 11” error
in osascript. Using `run script` does not segfault,
but no output is generated (as we had already noted:
`run script` was the first method I used to run the
tests). I have also tried to use
`do shell script “osascript …”`, and I have even tried
to call OSAKit directly, but the results are the
same as with `run script` (no output).

So, for the time being, the makefile can only be used
to build the tests into a script bundle. Then,
the tests must be run manually.