Allow specifying the encoding of source scripts.
This is done mainly to avoid:

    The file […] couldn’t be opened because the text encoding of its contents can’t be determined. (-2700)

Such an error happens, for instance, when using ASMake to compile
a MacRoman-encoded script (as `ASMake.applescript` is), but the encoding
is not present in the extended attributes of the file (which may occur
when cloning the Git repository).

Fixes https://github.com/lifepillar/ASMake/issues/6.